Проходят все тесты рецептов кроме фильтра

+++ backend/foodgram/food/models.py
@@ -41,7 +41,7 @@ class Ingredient(models.Model):
 class RecipeIngredient(models.Model):
     recipe = models.ForeignKey('Recipe', on_delete=models.CASCADE)
     ingredient = models.ForeignKey(Ingredient, on_delete=models.CASCADE)
-    amount = models.IntegerField(verbose_name='Количество')
+    amount = models.PositiveSmallIntegerField(verbose_name='Количество')
 
     class Meta:
         default_related_name = 'recipe_ingredients'
